chore: full audit — patch bytes/keccak CVEs, fix all clippy, add 35 tests, add cargo audit to CI, write README
Security:
- bytes -> 1.11.1 (RUSTSEC-2026-0007: integer overflow in BytesMut::reserve)
- sha3  -> 0.11.0 (pulls keccak 0.2.0, eliminating yanked keccak 0.1.5 / RUSTSEC-2026-0012)
- blake2 -> 0.10.6 (explicit pin)
- cargo audit: 0 vulnerabilities, 0 warnings

Clippy (all-targets, -D warnings):
- Fix empty_line_after_doc_comments in proto.rs (outer -> inner doc)
- Suppress large_enum_variant on generated waves.events include!
- Fix 10x &Vec<T> -> &[T] across repo/mod.rs and repo/pg.rs trait + impls
- Fix 2x std::i64::MAX -> i64::MAX in consumer/mod.rs
- Fix 1x std::i64::MAX -> i64::MAX + &Vec -> &[V] in repo/pg.rs
- Add type aliases for complex tuple types in insert_txs_16/18
- Rename UidGenerator::next -> next_uid (avoids confusion with Iterator::next)
- Fix redundant borrows flagged by -D warnings in tests

Tests:
- 35 tests pass (up from 3)
- New: utils.rs (into_base58, into_prefixed_base64, epoch_ms_to_naivedatetime, escape_unicode_null)
- New: waves.rs (is_valid_base58, extract_asset_id, is_dcc_asset_id, keccak256, blake2b256, Address)

CI:
- Add cargo-audit install + cargo audit --deny warnings step
- Runs after test step, blocks on any known vulnerability

README:
- Full DCC-specific documentation: architecture, env vars, Docker, development setup

+++ apps/blockchain-postgres-sync/src/lib/db.rs
@@ -35,7 +35,7 @@ pub fn pool(config: &Config) -> Result<PgPool, AppError> {
     let manager = ConnectionManager::<PgConnection>::new(db_url);
     Ok(Pool::builder()
         .min_idle(Some(1))
-        .max_size(config.poolsize as u32)
+        .max_size(config.poolsize)
         .idle_timeout(Some(Duration::from_secs(10 * 60)))
         .connection_timeout(Duration::from_secs(5))
         .build(manager)?)
